refactor integration, add tests and such. merge error_constants into constants
