+++ Tests/scripts/hook_validations/structure.py
@@ -13,8 +13,7 @@ import yaml
 from Tests.scripts.constants import YML_INTEGRATION_REGEXES, YML_SCRIPT_REGEXES, JSON_ALL_WIDGETS_REGEXES, \
     JSON_ALL_DASHBOARDS_REGEXES, JSON_ALL_CONNECTIONS_REGEXES, JSON_ALL_CLASSIFIER_REGEXES, \
     JSON_ALL_LAYOUT_REGEXES, JSON_ALL_INCIDENT_FIELD_REGEXES, YML_ALL_PLAYBOOKS_REGEX, JSON_ALL_REPORTS_REGEXES, \
-    MISC_REGEX, MISC_REPUTATIONS_REGEX
-from Tests.scripts.error_constants import Errors
+    MISC_REGEX, MISC_REPUTATIONS_REGEX, Errors
 from Tests.test_utils import print_error, get_matching_regex, get_remote_file
 
 try:
